webapp/ar-ctrgcn-webapp/requirements.txt
absl-py==1.0.0
attrs==21.4.0
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Flask==2.0.3
Flask-SocketIO==5.1.1
fonttools==4.31.1
h11==0.13.0
h5py==3.6.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.4.0
MarkupSafe==2.1.1
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyparsing==3.0.7
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
simple-websocket==0.5.1
six==1.16.0
threadpoolctl==3.1.0
#torch==1.11.0+cpu
#torchaudio==0.11.0+cpu
#torchvision==0.12.0+cpu
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wsproto==1.1.0
